refactor: move `cache-inline` files
