+++ docs/.vitepress/config.ts
@@ -7,7 +7,7 @@ import { withMermaid } from 'vitepress-plugin-mermaid'
 import { version } from '../../package.json'
 import { transformerColorizedBrackets } from '../../packages/colorized-brackets/src'
 import { transformerMetaWordHighlight, transformerNotationWordHighlight, transformerRemoveNotationEscape } from '../../packages/transformers/src'
-import { createTwoslashConfig } from '../../packages/vitepress-twoslash/src/config'
+import { createTwoslashConfig } from '../../packages/vitepress-twoslash/src/cache-inline'
 import { defaultHoverInfoProcessor, transformerTwoslash } from '../../packages/vitepress-twoslash/src/index'
 import vite from './vite.config'
